Fixed a naming error in heap header file

+++ src/heap.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include "headers/njvm.h"
 
-unsigned int heapSize;
+unsigned long heapSize;
 
 unsigned char* heap; /* Main Heap-Pointer */
 unsigned char* src;  /* Source pointer (Quell) - Unused*/
@@ -11,7 +11,7 @@ unsigned char* dest; /* Destination pointer (Ziel) - Used for allocation*/
 void initHeap(void) {
     heap = calloc(heapSize, sizeof(unsigned char));
     if (heap == NULL) {
-        printf("Error: System can't supply %u Bytes for heap creation!\n", heapSize);
+        printf("Error: System can't supply %lu Bytes for heap creation!\n", heapSize);
         exit(E_ERR_SYS_MEM);
     }
     dest = heap;
